Started on loading. Constants can be loaded as can the table, but not the table properties yet. More work to come.
Loading the tables caused an exception as previous gui components were not having their listeners removed. GUI components now implement WabitNode to cleanup their listeners on model components. 
All query model components now implement WabitObject as required for WabitNode.
The Section class was removed as it is not being used and removing it actually makes this change smaller.

This also fixes most if not all of the memory leaks coming from the GUI in the query side.