Overrode the JDBCDataSource's equals method (and subsequently, the hashcode method) so that it bases equality on the equality of the underlying SPDatasource. I don't use the underlying WabitObject name property because it is set to the same value as the SPDataSource's name.
This fixes the failing test case for testing persistence of JDBCDataSources, and I think this makes sense since we want to make sure the datasource and it's properties got persisted, and I don't think object reference equality makes sense.

A concern that I have here though is that the Wabit project file only saves the SPDataSource name, which unfortunately leads to the same problems that we have with Architect and MatchMaker, which is if we pass this file along to another person and they either have no data source with these names (they may have the actual database connection right but they named it differently), or they have data sources with the exact same name but the underlying connection details are different, then the whole project will very likely not work as desired. 

Ideally, we would store the underlying connection details along with the project (obfuscating passwords of course).