Deleting a datasource now prompts the user to either: 1)Delete that datasource along with all the queries and parts of layouts dependent on it. 2)Delete the datasource and set its queries to another datasource. Parts of the layouts built on these queries will be deleted. 3)Cancel(Forget about deleting the datasource)
Deleting a query will display a warning message that if the user proceeds with the delete, parts of layouts dependent on the query will be deleted as well.

On right-clicking a datasource in the project tree, it will display the project panel. This is a fix for a NPE thrown when deleting a datasource.