Now you specify the dialog owner for the server menu and action. This allows proper error message dialogs.
Also made good progress on the ability to actually create a session from a project on a remote server. There's a bug in the initialization (it tries to create a user prompter too early) but we believe it's very close to working. The next step is to create a new server-aware session context, which will get its data source list from the server.