Fixed the "magic rotating page" problem reported by Benjamin et al.
In support of this fix, we changed the way width, height, and orientation interact with each other in Page. (In other words, we changed the API of Page). Now the width and height measure the _apparent_ width and height of the page rather than the idealized width and height. This greatly simplifies use of pages that are already set up, but it does make creating a new page with landscape orientation a bit more of a head-scratcher than it used to be.

We also removed the almost-never-used StandardPageSizes enum (it was actually marked with an XXX tag asking to be removed). The replacement is the java print api PageFormat object.

Finally, in response to Sam's request, we have changed the default page orientation to landscape.

As a side effect of all this, the page is now much better at adjusting its margins when you switch it between portrait and landscape modes.