Fix for task: Ensure page orientation loads and saves properly
Also had to modify the session's hasUnsavedChanges() method because the old code was getting into an infinite loop when I was testing. The approach is still not correct, but the infinite loop isn't happening now. I created a new bug for the overhaul of the whole approach.

This change bumps the export file version number to 1.0.1.