Fixes bug 1917: Fix MDX query loading and saving when loading and saving project files http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1917
Also fixes bug 1934 :Loading a workspace with an OLAP query in a report layout throws an NPE 
http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1934

Delays OLAP query initialization as much as possible by caching the workspace XML data. There is still the issue where the last edited pane was related to OLAP; when loading back, the OLAP editor panel tries to init.