NEW - bug 1964: NullPointerException sometimes happens when adding a new datasource while in the OLAP query editor http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1964
The NewWorkspaceScreen was not removing a listener on the session when
the frame was being disposed. This led to the listener trying to
recreate the session's UI when a data source was added after the
window appeared. Since the UI was being recreated the current
CellSetViewer was losing its graphics object and causing a NPE.