NEW - bug 1976: Delete the active report http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1976
If a query, olap query, or layout is removed from the workspace and is
the object being edited in the workspace the object under edit will
become the workspace itself. This will cause the editor to change
from showing the object just removed from the project to the main
Wabit screen. Three tests have been added to test this for queries,
olap queries, and layouts.