NEW - bug 1987: Fix being able to insert new dimensions in any order http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1987
Fixed up some of the dimension ordering code. Also changed the behaviour of adding members that have the hierarchy already in the query. In this case, the members get added and don't replace the existing selections. While there's still a bit of behaviour to be fixed or to be discussed with the rest of the team, I believe this is closer to the behaviour that is desired by the consultants.