NEW - bug 2065: Add a break type to reports that suppress repeating values. http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2065
Modified how the result set renderer does the layout of a result set.
Now the ReportPositionRenderer creates the entire layout. This is done
by creating ResultSetCells for each value to be displayed in the
renderer as well as defining their layout. This was done to more easily
support adding groupings.