NEW - bug 2025: Add buttons for all standard functions http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2025
Added the 'megatree' toolbar with 'New', 'Open', and 'Save' buttons. Also added the icons for these actions. One thing left to do for this is that 'Save' needs to become a 'Save Everything' button. But don't call it that, the button would get too big ;)

Also, some of the menu items have icons and others don't, so ideally they either all should have icons or none of them should.