NEW - bug 2012: Drag to edit column widths http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2012
Added the ability to drag a column edge in a result set renderer to 
change the column width.

To support this the swing specific parts of the ReportContentRenderer 
were separated into their own classes. Now each class in the core
report package does not rely on Swing and there is a Swing compliment
to each core renderer.