NEW - bug 2026: Bigger buttons http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2026
Added some of the new 'big' icons for toolbar buttons. Still need to do the Query Pen which is in the Library, but will need to set it up so that the big icons only show up in Wabit. Also need to add text to the buttons for consistency with the MegaTree(TM) toolbar