NEW - bug 2020: A warning before executing a query with a missing join in the querypen http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2020
This commit has two changes in it that affect related files.
The first is splitting the swing worker registry. The registry was
previously in the context. Now each session is a swing worker registry
to track workers per session and one registry in the context to track
workers used in loading sessions before the sessions are created.

The other change involved in this commit is the first part of the work
on warning a user before executing a query with missing joins. The
dialog that is displayed needs to be fixed and the properties of the
query need to be saved.