ASSIGNED - bug 1848: Add some kind of indication that an OLAP Query is running http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1848
OLAP queries can now be run in the background. I still have to go over more calls to OlapQuery.execute() because there are some that are currently not running in the background but could/should be.

Finally, I have to make OlapQuery threadsafe. This shouldn't take too much doing, thankfully.