NEW - bug 2045: Fix resize behaviour of result panel in query pen http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2045
Set the lower split on the query panel to have a minimum size of 0 so 
the scroll panes it contains can be reduced to a smaller size and not
be cut off. If the user is new and does not have a preferred position
of the split panes the lower split pane will be defaulted to 3/4 of the
available panel size. This way the results will be visible when the 
query first executes.