NEW - bug 2132: Move MDX editor button into a toolbar like all the other editors http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2132
Moved the mdx editor execute button into a toolbar for consistency. It would be nice to add export and create chart/report buttons, but I've been thinking it would be best of we implement saving MDX query text first? Because otherwise, I don't think the mdx text gets persisted right now, so it may be hard to save them for use in a report or chart.