ASSIGNED - bug 2106: Make charts first class citizens http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2106
Fixed a bug in the way XY datasets are specified. Previously, the X axis selection wasn't "sticking" because of a data type mismatch between a JComboBox.setSelectedValue() and the types of values the combo box actually contained (ColumnIdentifier vs. String, respectively)

In order to understand what I was doing, I also renamed the DataTypeSeries enum to ColumnRole. I renamed references to it within ChartPanel and ColumnIdentifier but not elsewhere.