ASSIGNED - bug 2143: Make null content boxes work properly http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2143
Fixed how the delete key was not working when you individually click on a content box and click delete (instead of selecting it by dragging). The even manager for picoolo was being overridden causing those picoolo handled events to not go through. Furthermore fixed up a problem where the SwingRenderer would not be set if you loaded up a report.