NEW - bug 2139: Add a global preference to not automatically run queries http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2139
Changed the preference wording to be for disabling automatic execution, which I believe is closer to the intended behaviour. Also, I moved the constant into the WabitSessionContext, mainly because I don't believe that it's necessarily a Swing-specific preference, and also because when being referred to in outside classes, I don't think it's the best thing to be making static references to Impl classes.