NEW - bug 2142: NPE when creating report from empty query http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2142
Fixed an NPE when initializing the result set after receiving a null
result set. The result set renderer does not need to be initialized
if there is no result set as there will be no headers or columns to
initialize. Added two tests for executing and rendering a result set
that is null.