RESOLVED - bug 2016: The tree on the right hand side should be optional and therefore be in a dockable palette http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2016
Final step was to get the toolbar to stay above the whole split pane when the 
source component is docked. We accomplished this by adding a getToolBar() method
to WabitPanel, and arranging the whole thing from the session context according
to the current source component preference.

Since we had to extract the toolbar from every WabitPanel implementation anyway,
we took this opportunity to use the new WabitToolBarBuilder everywhere. This
actually resolves another bug too.