NEW - bug 2233: Can't copy connections http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2233
Dragging an individual data source from one session to another can now
be done.

Made some changes with this to the context to improve ease of testing.
The context can now be given a user prompter so it can run headless.
Added a test case for dropping objects onto a session.