NEW - bug 1842: Add OLAP sort functionality http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1842
Added pop-up menu sort items for all sort orders, and a clear sort item as well.

One thing I'm not sure if clear sort should be on the axis you clicked or on the opposite axis, because when you sort on a measure, it applies the sort on the opposite axis.