NEW - bug 2267: Encapsulate foreground/background execution policy in WabitSession http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2267
The first part of this bug has been completed. There is now a set of 
methods on each session that will execute a Runnable on the foreground
or a background thread. The events that are fired from WabitObjects
are now fired from the foreground thread.

WabitObjects that are currently making workers need to be changed to 
use the runInBackground method. We also need to force events fired from
the Query object in the library on the foreground thread.