NEW - bug 2267: Encapsulate foreground/background execution policy in WabitSession http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2267
The execute method on the ResultSetProducer now returns a collection of
result sets in a Future. If methods need to just execute the query they
can call execute and not use the results so they don't block. If a 
method needs to know what the result is before moving forward they can
get the results of the future which will cause the thread to block
until the results are available.