This code bomb is to add an addChild method to WabitObject. The addChild method will be useful for the server components and the undo manager.
The change to the library that removed the need to have Items as 
children of the Query reduces the complexity of the children of the
QueryCache. This also removes the need for the WabitQueryItem.

The addChild method has been implemented for the objects that have 
children but some addChild methods throw exceptions at current as their
children are managed internally and should not be modified by outside
classes.

Changes were made to the Wabit Olap objects to fix exceptions that were
being thrown if the query was accessed before it was initialized.