NEW - bug 2332: Make a refresh button on the Wabit client http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2332
In implementing this feature, I discovered that the existence
and use of WabitSessionContext.deregisterChildSession was quite
brittle. I've removed that method and implemented session
lifecycle listeners in the context implementations that need to
perform cleanup.

Additional note: I've noticed that the heap size does not go
down after a session has been closed by clicking its tab!
We will need to do a bit of memory profiling/analysis to find
out why!