Fixed the bug where a creating a Chart with duplicate columns caused an IOOBE.
Previously, this happened because Chart prunes out duplicate chart columns based on the names taken from a ResultSet. If it finds a match for the name, Chart does not add a new ChartColumn to its list of children. However, the CategoryChartHeaderRenderer uses the ResultSet to grab the index it needs to render, which is obviously inconsistent with the list of children Chart provides.

Now, each duplicate rendered column in the header shares the same ChartColumn. However, a separate instance of combo boxes for each column is created. When the role is changed in one of those duplicate combo boxes, the other combo boxes will update accordingly. In this setup, the duplicated rendered combo boxes can never differ in roles, they all either have to be none, category, or series.