Fixed issue where any kind of chart on a report would not load on a server project. The issue was caused by the report's cached chart.
The cached chart and the original chart shared the same chart columns. However, since these chart columns are children of the original chart, adding these columns to the cached chart caused their parent pointer to change. This is bad! New copies of these chart columns should be created instead so that they do not have any affect on the server.